application: move backend to application.backend
